Rewritten universal script
Todo: add manjaro function

+++ instalation-scripts/universal_ubuntu_install.sh
@@ -1,36 +1,34 @@
 #!/bin/bash
 # Installs top.sls and advanced-linux-apps
 # Check if script has been run before
-if [ -f /tmp/helloworld.txt ]; then
-    echo "===> This install script has already been run! It is intended to be run only once <==="
-    exit 0
-fi
 
 echo "Set to basic or advanced:"
 read MinionID
+if [ ! -f /etc/salt/minion ]; then
+    	echo "===> First run installing salt <==="
+   	sudo wget -O - https://repo.saltstack.com/apt/ubuntu/18.04/amd64/latest/SALTSTACK-GPG-KEY.pub | sudo apt-key add -
+	echo "deb [arch=amd64] http://repo.saltstack.com/apt/ubuntu/18.04/amd64/latest bionic main" | sudo tee /etc/apt/sources.list.d/saltstack.list
+	echo "="
+	echo "===> Updating packages... <==="
+	echo "="
+	sudo apt-get update && sudo apt upgrade -y
+	echo "="
+	echo "===> Installing git and salt... <==="
+	echo "="
+	sudo apt-get install git salt-minion -y 
+	echo "="
+	echo "===> Cloning repository... <==="
+	echo "="
+	cd /srv/
+	sudo git clone https://github.com/aksratamo/salt
+	cd salt/
 
-sudo wget -O - https://repo.saltstack.com/apt/ubuntu/18.04/amd64/latest/SALTSTACK-GPG-KEY.pub | sudo apt-key add -
-echo "deb [arch=amd64] http://repo.saltstack.com/apt/ubuntu/18.04/amd64/latest bionic main" | sudo tee /etc/apt/sources.list.d/saltstack.list
-echo "="
-echo "===> Updating packages... <==="
-echo "="
-sudo apt-get update && sudo apt upgrade -y
-echo "="
-echo "===> Installing git and salt... <==="
-echo "="
-sudo apt-get install git salt-minion -y 
 
-echo "Writing salt settings to file and restarting salt-minion..."
-echo -e "master: localhost\nid: linux-$MinionID" | sudo tee /etc/salt/minion
+fi
 
-echo "="
-echo "===> Cloning repository... <==="
-echo "="
-cd /srv/
-sudo git clone https://github.com/aksratamo/salt
-cd salt/
-
-echo "="
-echo "===> Running highstate and advanced-linux, Please wait <==="
-echo "="
-sudo salt-call --local state.apply 
+    	echo "===> Just changing Minion ID <==="
+	echo -e "master: localhost\nid: linux-$MinionID" | sudo tee /etc/salt/minion
+	echo "="
+	echo "===> Running highstate with MinionID ¤ $MinionID ¤ , Please wait <==="
+	echo "="
+	sudo salt-call --local state.apply
